requirements.txt
typer[all]
rich
questionary
requests
python-dotenv
